Becoming LORDS OF C

+++ 0x00-hello_world/6-size.c
@@ -8,7 +8,7 @@ int main(void)
 	long long int l;
 	float f;
 
-	printf("Size of a char: %i byte(s)\n", (unsigned long)sizeof(x));
+	printf("Size of a char: %i byte(s)\n", sizeof(x));
 	printf("Size of an int: %i byte(s)\n", sizeof(n));
 	printf("Size of a long int: %i byte(s)\n", sizeof(d));
 	printf("Size of a long long int: %i byte(s)\n", sizeof(l));
